Revert "add Hero to prayer image"
This reverts commit 14e251723a4c4df113dffd5c7f2a4db965d2dc8f.

This should be done with a different approach as the Hero is not in the widget tree when the FutureBuilder is loading, so the transition does not work. Moving up the Hero to be the parent of FutureBuilder doesn't look good neither, as the transition back shows the progress indicator for a short time as it was initially contained by Hero widget.